+++ src/main/java/woowacourse/shoppingcart/application/dto/ProductResponse.java
@@ -1,6 +1,6 @@
 package woowacourse.shoppingcart.application.dto;
 
-import woowacourse.shoppingcart.domain.Product;
+import woowacourse.shoppingcart.domain.product.Product;
 
 public class ProductResponse {
