+++ src/test/java/woowacourse/shoppingcart/acceptance/OrderAcceptanceTest.java
@@ -19,7 +19,7 @@ import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import woowacourse.shoppingcart.domain.Orders;
+import woowacourse.shoppingcart.domain.order.Orders;
 import woowacourse.shoppingcart.ui.dto.OrderRequest;
 
 @DisplayName("주문 관련 기능")
