+++ src/test/java/woowacourse/shoppingcart/dao/OrdersDetailDaoTest.java
@@ -12,7 +12,7 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.jdbc.Sql;
-import woowacourse.shoppingcart.domain.OrderDetail;
+import woowacourse.shoppingcart.domain.order.OrderDetail;
 
 @JdbcTest
 @AutoConfigureTestDatabase(replace = Replace.NONE)
