+++ src/test/java/woowacourse/shoppingcart/ui/OrderControllerTest.java
@@ -23,9 +23,9 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 import woowacourse.shoppingcart.application.OrderService;
-import woowacourse.shoppingcart.domain.OrderDetail;
-import woowacourse.shoppingcart.domain.Orders;
-import woowacourse.shoppingcart.domain.Product;
+import woowacourse.shoppingcart.domain.order.OrderDetail;
+import woowacourse.shoppingcart.domain.order.Orders;
+import woowacourse.shoppingcart.domain.product.Product;
 import woowacourse.shoppingcart.ui.dto.OrderRequest;
 
 @SpringBootTest
